requirements.txt
hstrader
pandas
plotly
python-dotenv
numpy
dash